third_party/blink/web_tests/platform/generic/fast/events/touch/touch-over-hidden-iframe-expected.txt

Touches go here

Tests that touch events on a an element covered by a hidden iframe trigger the touch handlers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS touch.pageX is 50
PASS touch.pageY is 50
PASS successfullyParsed is true

TEST COMPLETE
